refactor: rename components folder
