+++ src/view/FooterView/Alpine/FooterViewAlpine.ts
