+++ src/view/FooterView/FooterView.test.ts
