+++ src/view/FooterView/FooterView.ts
